[server] Move remaining trivial RPC away from EE
